Separate parsers into modules
