+++ siri2gtfsrt.py
@@ -1,20 +1,19 @@
 #!/usr/bin/python
 
-import json
 import gtfs_realtime_pb2
 from google.protobuf import text_format
 from flask import Flask
-from flask import request, abort
+from flask import request
 from urllib2 import urlopen
-import time
-import datetime
-import dateutil.parser
-import pytz
 import os
 import threading
 from traceback import print_exc
 
-#Tampere realtinme siri feed
+import hsl
+import foli
+import gtfs
+
+# Tampere realtinme siri feed
 JOLI_URL = os.environ.get('JOLI_URL', "http://data.itsfactory.fi/journeys/api/1/vehicle-activity")
 # HSL realtime siri feed
 HSL_URL = os.environ.get('HSL_URL', "http://api.digitransit.fi/realtime/navigator-server/v1/siriaccess/vm/json?operatorRef=HSL")
@@ -23,7 +22,6 @@ TRAIN_URL = os.environ.get('TRAIN_URL', "http://api.digitransit.fi/realtime/rail
 # HSL area service-alerts GTFS-RT feed
 TRIP_UPDATE_URL = os.environ.get('TRIP_UPDATE_URL', "http://api.digitransit.fi/realtime/service-alerts/v1/")
 
-EPOCH = datetime.datetime(1970, 1, 1, tzinfo=pytz.utc)
 
 #global data
 ctx = {"msg" : gtfs_realtime_pb2.FeedMessage()}
@@ -60,66 +58,6 @@ class Poll(object):
 
             self.stopped.wait(self.interval)
 
-def handle_siri(raw):
-    siri_data = json.loads(raw.decode('utf-8'))['Siri']
-    msg = gtfs_realtime_pb2.FeedMessage()
-    msg.header.gtfs_realtime_version = "1.0"
-    msg.header.incrementality = msg.header.FULL_DATASET
-    msg.header.timestamp = int(siri_data['ServiceDelivery']['ResponseTimestamp']) / 1000
-
-    for i, vehicle in enumerate(siri_data['ServiceDelivery']['VehicleMonitoringDelivery'][0]['VehicleActivity']):
-        route_id = vehicle['MonitoredVehicleJourney']['LineRef']['value'][:5].strip()
-
-        if route_id in ('1300', '1300V', '1300M' ):
-            continue # No other information than location for metros
-
-        if route_id[0:4] in ('3001', '3002'):
-            continue # Train data is better at rata.digitraffic.fi
-
-        if route_id[0] in ('k', 'K'):
-            continue # Kutsuplus
-
-        if 'Delay' not in vehicle['MonitoredVehicleJourney']:
-            continue
-
-
-        ent = msg.entity.add()
-        ent.id = str(i)
-        ent.trip_update.timestamp = vehicle['RecordedAtTime']/1000
-        ent.trip_update.trip.route_id = route_id
-
-        try:
-            int(vehicle['MonitoredVehicleJourney']['Delay'])
-        except:
-            print_exc()
-            print vehicle, vehicle['MonitoredVehicleJourney']['Delay']
-
-            continue
-
-        ent.trip_update.trip.start_date = vehicle['MonitoredVehicleJourney']['FramedVehicleJourneyRef']['DataFrameRef']['value'].replace("-","")
-        if 'DatedVehicleJourneyRef' in vehicle['MonitoredVehicleJourney']['FramedVehicleJourneyRef']:
-            start_time = vehicle['MonitoredVehicleJourney']['FramedVehicleJourneyRef']['DatedVehicleJourneyRef']
-            ent.trip_update.trip.start_time = start_time[:2]+":"+start_time[2:]+":00"
-
-        if 'DirectionRef' in vehicle['MonitoredVehicleJourney'] and 'value' in vehicle['MonitoredVehicleJourney']['DirectionRef']:
-            ent.trip_update.trip.direction_id = int(vehicle['MonitoredVehicleJourney']['DirectionRef']['value'])-1
-
-        if 'VehicleRef' in vehicle['MonitoredVehicleJourney']:
-            ent.trip_update.vehicle.label = vehicle['MonitoredVehicleJourney']['VehicleRef']['value']
-
-        stoptime = ent.trip_update.stop_time_update.add()
-
-        if 'MonitoredCall' in vehicle['MonitoredVehicleJourney']:
-            if 'StopPointRef' in vehicle['MonitoredVehicleJourney']['MonitoredCall']:
-                stoptime.stop_id = vehicle['MonitoredVehicleJourney']['MonitoredCall']['StopPointRef']
-            elif 'Order' in vehicle['MonitoredVehicleJourney']['MonitoredCall']:
-                stoptime.stop_sequence = vehicle['MonitoredVehicleJourney']['MonitoredCall']['Order']
-            stoptime.arrival.delay = int(vehicle['MonitoredVehicleJourney']['Delay'])
-        else:
-            ent.trip_update.delay = int(vehicle['MonitoredVehicleJourney']['Delay'])
-
-    return msg
-
 def handle_trip_update(orig_msg, alerts):
     if alerts == None:
         return
@@ -128,52 +66,6 @@ def handle_trip_update(orig_msg, alerts):
         if entity.HasField('trip_update'):
             new_entity = orig_msg.entity.add()
             new_entity.CopyFrom(entity)
-
-def parse_gtfsrt(raw):
-    if raw == None:
-        return None
-    msg = gtfs_realtime_pb2.FeedMessage()
-    msg.ParseFromString(raw)
-
-    return msg
-
-def handle_journeys(raw):
-    journeys_data = json.loads(raw.decode('utf-8'))
-    if journeys_data['status'] != "success":
-        abort(500)
-    msg = gtfs_realtime_pb2.FeedMessage()
-    msg.header.gtfs_realtime_version = "1.0"
-    msg.header.incrementality = msg.header.FULL_DATASET
-    msg.header.timestamp = int(time.time())
-
-    for i, vehicle in enumerate(journeys_data['body']):
-        ent = msg.entity.add()
-        ent.id = str(i)
-
-        route_id = vehicle['monitoredVehicleJourney']['lineRef']
-        ent.trip_update.trip.route_id = route_id
-
-        date = vehicle['monitoredVehicleJourney']['framedVehicleJourneyRef']['dateFrameRef'].replace("-","")
-        ent.trip_update.trip.start_date = date
-
-        start_time = vehicle['monitoredVehicleJourney']['originAimedDepartureTime']
-        ent.trip_update.trip.start_time = start_time[:2]+":"+start_time[2:]+":00"
-
-        direction = vehicle['monitoredVehicleJourney']['directionRef']
-        ent.trip_update.trip.direction_id = int(direction)-1
-
-
-        if 'onwardCalls' not in vehicle['monitoredVehicleJourney']:
-            continue
-
-        for call in vehicle['monitoredVehicleJourney']['onwardCalls']:
-            stoptime = ent.trip_update.stop_time_update.add()
-            stoptime.stop_sequence = int(call['order'])
-            arrival_time = (dateutil.parser.parse(call['expectedArrivalTime']) - EPOCH).total_seconds()
-            stoptime.arrival.time = int(arrival_time)
-            departure_time = (dateutil.parser.parse(call['expectedDepartureTime']) - EPOCH).total_seconds()
-            stoptime.departure.time = int(departure_time)
-    return msg
 
 # data is processed asynchronously as new data come in
 def process_hsl_data():
@@ -199,10 +91,10 @@ def process_hsl_data():
 
     ctx["msg"] = ctx["nmsg"];
 
-HSL_poll = Poll(HSL_URL, 60, handle_siri, True)
-JOLI_poll = Poll(JOLI_URL, 60, handle_journeys, False)
-TRAIN_poll = Poll(TRAIN_URL, 60, parse_gtfsrt, True)
-TRIP_UPDATE_poll = Poll(TRIP_UPDATE_URL, 60, parse_gtfsrt, True)
+HSL_poll = Poll(HSL_URL, 60, hsl.handle_siri, True)
+JOLI_poll = Poll(JOLI_URL, 60, foli.handle_journeys, False)
+TRAIN_poll = Poll(TRAIN_URL, 60, gtfs.parse_gtfsrt, True)
+TRIP_UPDATE_poll = Poll(TRIP_UPDATE_URL, 60, gtfs.parse_gtfsrt, True)
 
 app = Flask(__name__)
